+++ packages/ketchup/src/components/kup-data-table/readme.md
@@ -167,6 +167,7 @@ Type: `Promise<void>`
 - [kup-radio](../kup-radio)
 - [kup-paginator](../kup-paginator)
 - [kup-combobox](../kup-combobox)
+- [kup-badge](../kup-badge)
 
 ### Graph
 ```mermaid
@@ -186,6 +187,7 @@ graph TD;
   kup-data-table --> kup-radio
   kup-data-table --> kup-paginator
   kup-data-table --> kup-combobox
+  kup-data-table --> kup-badge
   kup-date-picker --> kup-text-field
   kup-date-picker --> kup-button
   kup-tooltip --> kup-button
